services/executor-python/requirements.txt
pyzmq==26.2.0
protobuf==5.29.0
ib-insync==0.9.86
